Extract kernel name normalization for torch trace coverage tests.
Move normalize_kernel_name helpers into torch_trace_kernel_names and
register the unit test as test_torch_trace_kernel_names.
