+++ projects/rocprofiler-compute/tests/test_categories.yaml
@@ -55,7 +55,7 @@ test_categories:
       - test_csv_compression
       - test_inject_roctx_package
       - test_torch_cpp_loader
-      - test_torch_trace_coverage_utils
+      - test_torch_trace_kernel_names
       - test_conftest
       - test_mem_chart_gfx9
       - test_mem_chart_gfx11
@@ -142,7 +142,7 @@ test_categories:
       - test_csv_compression
       - test_inject_roctx_package
       - test_torch_cpp_loader
-      - test_torch_trace_coverage_utils
+      - test_torch_trace_kernel_names
       - test_profiler_base
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
@@ -235,7 +235,7 @@ test_categories:
       - test_csv_compression
       - test_inject_roctx_package
       - test_torch_cpp_loader
-      - test_torch_trace_coverage_utils
+      - test_torch_trace_kernel_names
       - test_profiler_base
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
@@ -328,7 +328,7 @@ test_categories:
       - test_csv_compression
       - test_inject_roctx_package
       - test_torch_cpp_loader
-      - test_torch_trace_coverage_utils
+      - test_torch_trace_kernel_names
       - test_profiler_base
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
